# public/docs/openapi.yaml
openapi: 3.0.0
info:
  title: 'Cairo Store API'
  description: "كل نقاط المتجر ولوحة التحكم.\r\n\r\nملاحظة عامة على الأخطاء: نقاط JSON في هذا المشروع تُرجع 200 حتى عند\r\nفشل التحقق، والنتيجة تُقرأ من الحقل success في الجسم لا من كود HTTP.\r\nالاستثناءات الوحيدة هي 302 (تحويل لتسجيل الدخول) و403 (صلاحية ناقصة)\r\nو404 (راوت غير موجود). هذا سلوك قائم يعتمد عليه الـfrontend، ووُثّق\r\nهنا كما هو لا كما ينبغي أن يكون."
  version: 1.1.0
servers:
  -
    url: 'http://localhost/STORE/public'
    description: 'Local / Development Server'
paths:
  /admin/settings:
    get:
      tags:
        - 'Admin - Site Settings'
      summary: 'Show the current site settings page'
      operationId: 634d1801e5ae45f856fbfc627b8bcd25
      responses:
        '200':
          description: 'HTML page — requires the can_edit_site_content permission'
        '302':
          description: 'Redirect to /admin/login'
        '403':
          description: 'Forbidden — the admin lacks can_edit_site_content'
      security:
        -
          adminSessionAuth: []
    post:
      tags:
        - 'Admin - Site Settings'
      summary: 'Save site settings (AJAX) — the financial and currency fields are accepted only from a holder of can_manage_checkout_settings'
      description: "AdminSiteSettingsController — the admin's site settings page.\nExtends AdminController, which verifies the admin login automatically."
      operationId: 3e053842fbd06f520e821e6726b6fe5f
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              required:
                - csrf_token
              properties:
                csrf_token:
                  type: string
                footer_text:
                  type: string
                facebook_url:
                  type: string
                instagram_url:
                  type: string
                snapchat_url:
                  type: string
                whatsapp_number:
                  type: string
                tiktok_url:
                  type: string
                twitter_x_url:
                  type: string
                google_maps_url:
                  type: string
                copyright_text:
                  type: string
                phone_number:
                  type: string
                working_hours:
                  type: string
                employees_count:
                  type: integer
                site_url:
                  type: string
                return_policy:
                  type: string
                privacy_policy:
                  type: string
                terms_and_conditions:
                  type: string
                default_currency:
                  description: 'Only for a holder of can_manage_checkout_settings'
                  type: string
                default_language:
                  description: 'Only for a holder of can_manage_checkout_settings'
                  type: string
              type: object
      responses:
        '200':
          description: 'Whether the save succeeded'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  message: { type: string }
                type: object
      security:
        -
          adminSessionAuth: []
...
  /admin/support/reply:
    post:
      tags:
        - 'Admin - Support'
      summary: 'Reply to a support message, delivered as a notification to its author'
      description: "AdminSupportController — the admin's support inbox page.\nExtends AdminController, which verifies the admin login automatically."
      operationId: 1352770916a2a4c08bb0492de4740ed7
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              required:
                - csrf_token
                - user_id
                - reply_text
              properties:
                csrf_token:
                  type: string
                user_id:
                  type: integer
                reply_text:
                  type: string
              type: object
      responses:
        '200':
          description: 'Whether sending succeeded'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  message: { type: string }
                type: object
      security:
        -
          adminSessionAuth: []
  /admin/support/delete:
    post:
      tags:
        - 'Admin - Support'
      summary: 'Permanently delete a support message'
      operationId: ff32f904c4a196d8bea9075d40a781bf
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              required:
                - csrf_token
                - message_id
              properties:
                csrf_token:
                  type: string
                message_id:
                  type: integer
              type: object
      responses:
        '200':
          description: 'Whether the deletion succeeded'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  message: { type: string }
                type: object
      security:
        -
          adminSessionAuth: []
  /admin/support:
    get:
      tags:
        - 'Admin - Support'
      summary: 'Support message list (search + pagination); opening it marks every message read automatically'
      operationId: 505ebd03f2ff82b6f81567650b77e89c
      parameters:
        -
          name: q
          in: query
          description: 'Search by name, email or message body'
          required: false
          schema:
            type: string
        -
          name: page
          in: query
          description: 'Page number'
          required: false
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: 'HTML page — requires the can_manage_support permission'
        '302':
          description: 'Redirect to /admin/login'
        '403':
          description: 'Forbidden — the admin lacks can_manage_support'
      security:
        -
          adminSessionAuth: []
...
components:
...
  securitySchemes:
    adminSessionAuth:
      type: apiKey
      description: 'PHP session cookie — يُنشأ تلقائياً عند تسجيل الدخول ويُرسل مع كل طلب محمي'
      name: admin_session
      in: cookie
    userSessionAuth:
      type: apiKey
      description: "جلسة المستخدم العادي — منفصلة تماماً عن admin_session بالاسم\r\nوالمحتوى، فلا يمكن الوصول لبيانات الأدمن من جلسة مستخدم ولا العكس\r\n(راجع isUser/isAdmin في auth_helper.php)."
      name: PHPSESSID
      in: cookie
...